Move Audit model to core app and update imports
