+++ core/models.py
@@ -1,0 +1,10 @@
+from django.db import models
+
+# Create your models here.
+class Audit(models.Model):
+    created_at = models.DateTimeField(auto_now_add=True)
+    updated_at = models.DateTimeField(auto_now=True)
+    deleted_at = models.DateTimeField(null=True, blank=True)
+
+    class Meta:
+        abstract = True
